Added Component Level Get/Set API
Now it is possible to get and set individual color space components without having to deal with Arrays. Conversion of spaces only take place if the color space does not match with the current ('this') chroma object.

For example, if the current chroma object is ChromaRGB type and the method setRGB_R(255); directly updates the ChromaRGB's rgb_R field.

If instead, setHSL_H(180); is called, then the chroma object is converted from ChromaRGB type to ChromaHSL type. The factory pattern handles the conversion process.