Added Channel Enum for Get/Set Methods
Now it is possible to use get/set methods for Chroma colors using Channel enumerables. It cleans up the method names appended with channels. Channels (such as R, G, B) can now be specified in the get/set method argument. Set methods are WIP and will be implemented in the next commit.