Added saturateBy() function
closes #3

`saturateBy(amount)` can be used to push Chroma color by a specific level of intensity. The function will make sure that the Chroma color is not clipped in the CIE color space. If the requested chroma amount is larger than the maximum chroma (as returned by the `getMaxChroma()` function), the returned intensity is clipped at the max chroma of the color. This function is useful when adding a fixed amount of intensity to a color.